Render react modals in a specific HTML element
