+++ app/frontend/js/apps/ExtractionApp/components/ParameterDeleteModal.jsx
@@ -1,0 +1,50 @@
+import React from 'react';
+import { createPortal } from 'react-dom';
+import Button from "react-bootstrap/Button";
+import Modal from "react-bootstrap/Modal";
+import { useSelector, useDispatch } from "react-redux";
+
+import { deleteParameter } from "~/js/features/ExtractionApp/ParametersSlice";
+import { selectAppDetails } from "~/js/features/ExtractionApp/AppDetailsSlice";
+import { selectUiAppDetails } from "~/js/features/ExtractionApp/UiAppDetailsSlice";
+
+const ParameterDeleteModal = ({ showModal, handleClose, id, name }) => {
+  const dispatch = useDispatch();  
+  const appDetails = useSelector(selectAppDetails);
+  const uiAppDetails = useSelector(selectUiAppDetails);
+  
+  const handleDeleteClick = () => {
+    dispatch(
+      deleteParameter({
+        id: id,
+        harvestDefinitionId: appDetails.harvestDefinition.id,
+        pipelineId: appDetails.pipeline.id,
+        extractionDefinitionId: appDetails.extractionDefinition.id,
+        requestId: uiAppDetails.activeRequest,
+      })
+    );
+    handleClose();
+  };
+
+  return createPortal(
+    <Modal show={showModal} onHide={handleClose}>
+      <Modal.Header closeButton>
+        <Modal.Title>Delete</Modal.Title>
+      </Modal.Header>
+      <Modal.Body>
+        Are you sure you want to delete the parameter "{name}"?
+      </Modal.Body>
+      <Modal.Footer>
+        <Button variant="secondary" onClick={handleClose}>
+          Close
+        </Button>
+        <Button variant="danger" onClick={handleDeleteClick}>
+          Delete
+        </Button>
+      </Modal.Footer>
+    </Modal>,
+    document.getElementById('react-modals')
+  )
+}
+
+export default ParameterDeleteModal;
