+++ app/views/layouts/application.html.erb
@@ -29,5 +29,6 @@
       <%= yield %>
     </div>
 
+    <span id='react-modals'></span>
   </body>
 </html>
